phoenix-bios-file-length-handling: Newer Phoenix BIOSes can have length > 1MB
Some newer Phoenix BIOSes, mostly from Supermicro such as those for motherboards
X7DA8 and X7DB8 now have BIOS length > 1024KB. This leads to failure to extract
since the signature check in the PhoenixModule() function fails. We adjust the
length based on the BIOSOffset parameter and the extraction works.

Change-Id: I275fc7904fb4e1163c9f68bdff9032431501a37d
Signed-off-by: Vikas N Kumar <vikas@cpan.org>
Reviewed-on: http://review.coreboot.org/1829
Tested-by: build bot (Jenkins)
Reviewed-by: Patrick Georgi <patrick@georgi-clan.de>