phoenix: don't break searching for PhoenixIDs too early
The loop over struct PhoenixIDs was broken after it found either a
BCPSYS id or a BCPFFV block. If a bios image contains both, both should
be interpreted. In fact PhoenixFFV() further down in the function is
only called when both were found.

This change makes it possible to extract the bios image contained in the
Flash BIOS Update Boot Disk for Fujitsu Siemens CELSIUS R540, Version
6.10.1.09.

Change-Id: Ia810e31af8e6a11af01bc30027675b5899f38229
Fixes: b09d0ef801b8 ("Add Phoenix SecureCore support")
Signed-off-by: Uwe Kleine-König <uwe@kleine-koenig.org>
Reviewed-on: http://review.coreboot.org/10479
Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Tested-by: build bot (Jenkins)
Reviewed-by: Patrick Georgi <pgeorgi@google.com>