ami: add some more image types
These were seen in the wild. Some of the already codes are apparently freely
used by the OEMs and thus may collide (e.g. the Asrock images, the SLIC
and perhaps the CompuTrace and such). It's not clear what to do about that.
Perhaps just keep them as they are, they're really just to be used as a hint
anyway.

Change-Id: Ic5a392ea8aa4b80090a308da58fe52f1f04ca83a
Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>
Reviewed-on: https://review.coreboot.org/20786
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Jonathan Neuschäfer <j.neuschaefer@gmx.net>
Reviewed-by: Martin Roth <martinroth@google.com>