dell_inspiron_1100_unpacker: Add module type 0x52
The module with type 0x52 in the Latitude E6400 BIOS is the EC firmware,
and is not compressed. The current behavior tries to decompress this
module leading to an index out of bounds error at runtime, so add a new
condition to avoid decompressing it. This assumes type 0x52 is always
used for EC firmware.

The current behavior also assumes all the data before the first module
is EC firmware, which was probably true for the Inspiron 1100, but for
the E6400 the EC firmware is in module 0x52. However, for simplicity an
exception is not made for this behavior, so the extracted EC.bin should
just be treated as spurious data for the E6400.

TEST: Inspiron 1100 BIOS is still unpacked correctly, Latitude E6400
BIOS unpacks without runtime errors.

Signed-off-by: Nicholas Chin <nic.c3.14@gmail.com>
Change-Id: I3152150b7dea4d79840c61683692c65b1311cce2
Reviewed-on: https://review.coreboot.org/c/bios_extract/+/74977
Reviewed-by: Paul Menzel <paulepanter@mailbox.org>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>